Add files via upload
After Rasool was able to get 100 in the Sort question of exercise 3, now he wants to solve the same question using function to challenge himself in the topic of function. The difference of this question is that the number of input numbers is not specified and the end of the numbers is determined by entering the number 0 by the user. Also, instead of printing all the numbers, only the i-th number after the Sort operation that the user enters in the input needs to be printed. has it.